Updated animation, speed does not depend on FPS

+++ TerraMU/AnimationPendulum.h
@@ -1,9 +1,10 @@
 #pragma once
 #include "Animation.h"
+#include "Maths.h"
 
 class AnimationPendulum : public Animation {
 public:
-	AnimationPendulum(float numberOfTimes, int period, vec2 position, vec2 scale, float offset) :
+	AnimationPendulum(float numberOfTimes, float period, vec2 position, vec2 scale, float offset) :
 		Animation(numberOfTimes, period, position, scale, offset) {};
 
 	mat3 next();
